Fix hashing output, parallelise creating generics, remove websockets
Refactor the hashing output

Remove Websockets as it was causing too many issues

The websockets crate contains 2 vulnerabilities which
I lack the time to fix. As well as this windows appears
to be stopping websockets from working. Looking into
replacing this with a more up to date library. Or using
hyper/similar directly

Use rayon to parallelise the creation of generics

Because we spend a huge amount of time generating the profile
data for generics we can do it in parallel. This works well because
from testing roughly 60% of the time is spent waiting for mediainfo
to start

Correctly track remaining hashes to be calculated