Reimplement websockets using tokio-tungstenite
The below were squashed and merged

Add a config setting for the port that the websocket will listen on
as well as a flag to overwrite the config

Handle failure to bind to one or both ports

Update some dependencies, turn off some default features

Listen for websockets on ipv4 and ipv6

After looking into it tokio-tungstenite seems like the
easier choice and has a ctrl-c listener built in

Add ability to add tasks via web socket, fix regression in manager

The web socket thread now acceps the commands process, import, and hash
they add the corresponding task to the end of the queue as a
proof of concept.

The regression was caused by me in b6b406e9be735d410df24587232a5dc770b55f79
making process_new files parallel meant that I switched a for while loop
that popped elements to a for loop. This meant that new_files_queue
was no longer being emptied and duplicates would be created if you
processed files more than one. This is fixed by simply clearing
new_files_queue after the for loop