Split hash function, rust_fmt. Related issue #26.
There is a tiny fix to model.rs in here as well where an option was
unwrapped and then checked
Hash function is now seperated into two different functions.
generic.hash and generic.fast_hash achieve the same thing as before but
now they go through an intermediate function that gets a hash from a
PathBuf. This will allow us to verify hashes against cached files for
the workers to ensure that no loss of data occured during caching