Add WorkerTranscodeQueue, store handle for transcode Command process
Add basic control over where the encode task goes into the worker queue. Might need to rethink how the workers queue is mirrored server-side.

The WorkerTranscodeQueue contains all controls over manipulating the actual queue, this will need to be tied into the WorkerManager, time to put my thinking cap on.

The handle for the currently running transcode is stored, it is used for both waiting for output from the process (which is handled in run_transcode() and for killing the process prematurely because of forced work being added.