Switch from RID to UID
A UID is handed out by the server on initialisation if the worker doesn't already have one in their config, the new UID is then written to the config file.

Removed generate_uid()

The UID has been switched from a String to a usize

Currently, the server will hand out duplicate UID's when a new worker is created, if it is created in another session.